requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary>=2.9.3
django-ckeditor==6.7.1
python-dotenv==1.0.1
